Prevents NullReferenceException on Xamarin iOS
When getting the value of a key that’s not present in the JSON, Xamarin
iOS throws a NullReferenceException when T is a value type. The null
check at the start of Extensions.Convert<JToken, T>(token) fails
despite passing a null value in. May be a Xamarin bug or a limitation
of Mono generics running on iOS.

Workaround performs the null check inside JToken.Value method where the
Mono runtime correctly recognises the token as null.